backend/ai_suggestions/agents_guider/pyproject.toml
[project]
name = "agents-guider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.9",
    "python-dotenv>=1.2.1",
]
